bioacoustics/ingestion/requirements.txt
grpcio==1.53.0
milvus[client]==2.2.8
tensorflow==2.12.0
faiss-cpu==1.7.4
google-cloud-storage==2.10.0